Planet: Change data type long -> float
git-svn-id: https://codex.cs.bham.ac.uk/svn/team-project/D2@110 cfd7d107-c69a-4b07-9ecf-543bb1cfc744

+++ module-planets/src/d2/teamproject/module/planets/Planet.java
@@ -6,7 +6,7 @@ import java.util.Comparator;
 
 public class Planet implements Comparable<Planet> {
     private String name;
-    private long mass, diameter, distToSun;
+    private float mass, diameter, distToSun;
     private Comparator<Planet> cmp;
 
     @Override
@@ -18,15 +18,15 @@ public class Planet implements Comparable<Planet> {
         return name;
     }
 
-    public long getMass() {
+    public float getMass() {
         return mass;
     }
 
-    public long getDiameter() {
+    public float getDiameter() {
         return diameter;
     }
 
-    public long getDistToSun() {
+    public float getDistToSun() {
         return distToSun;
     }
